requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna==2.9
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-docx==0.8.10
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
sly==0.4
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
